feat(builder): embedded AI assistant panel scoped to the open resume (#27)
- extract the agent chat (messages, streaming, patch cards, composer) into a shared
  AgentChat component reused by both the /agent route and the builder
- add a Sparkle toggle in the builder header opening the chat as a right-side Sheet
  (full-screen on mobile); gated on a usable AI provider with a setup hint otherwise
- add agent.threads.getOrCreateForResume: an in-place thread whose working resume IS the
  open resume, so the agent's server-side patches flow back live via the existing builder
  update subscription (and are snapshotted as "AI edit" by the version history)

Verified in-browser: the toggle opens the panel; with no provider it shows the setup hint.
